replacing the old filter component
